Split driver and node into 2 files
