+++ neato_description/launch/description.launch.py
@@ -49,6 +49,5 @@ def generate_launch_description():
           executable='robot_state_publisher',
           name='robot_state_publisher',
           output='screen',
-          parameters=[{'use_sim_time': use_sim_time}],
-          arguments=[urdf_params]),
+          parameters=[{'use_sim_time': use_sim_time, 'robot_description': robot_description}]),
   ])
